requirements.txt
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
Werkzeug==3.0.1
redis==5.0.1
uagents>=0.9.0
fetchai-ledger-api>=0.3.0
boto3>=1.26.0
requests>=2.28.0
prometheus-flask-exporter>=0.18.2
prometheus-client>=0.9.0
PyYAML>=6.0.1
